test(components): Moves mocks into stories dir
