+++ src/stories/Icons.stories.tsx
@@ -12,7 +12,6 @@ import {
 	ArrowsDownIcon,
 	ArrowsDownUpIcon,
 	ArrowsUpIcon,
-	Avatar,
 	ClipboardIcon,
 	CloseIcon,
 	CopyIcon,
@@ -54,38 +53,7 @@ import { ICON_WEIGHTS } from "@consts";
 import type { IconWeight } from "@phosphor-icons/react";
 import type { Meta, StoryObj } from "@storybook/react";
 import { expect, within } from "@storybook/test";
-import { imageSrcMocks, weightControl } from "@utils";
-
-export const iconControl = {
-	control: "radio",
-	options: [
-		"Nothing",
-		ArrowLineDownIcon.displayName,
-		DefaultIcon.displayName,
-		FourLeafCloverIcon.displayName,
-	],
-	mapping: {
-		ArrowLineDownIcon,
-		DefaultIcon,
-		FourLeafCloverIcon,
-		Nothing: undefined,
-	},
-};
-
-export const imageMocks = {
-	woman: <img alt="woman" width={24} height={24} src={imageSrcMocks.woman} />,
-	flowerIcon: <FourLeafCloverIcon size={16} alt="flower icon" />,
-	avatar: <Avatar img={imageSrcMocks.man} username="" />,
-};
-
-export const imageControl = {
-	control: "radio",
-	options: ["woman", "flowerIcon", "avatar", "Nothing"],
-	mapping: {
-		...imageMocks,
-		Nothing: null,
-	},
-};
+import { weightControl } from "./mocks";
 
 const allIcons = {
 	AddIcon,
@@ -173,6 +141,15 @@ const meta = {
 
 export default meta;
 type Story = StoryObj<typeof meta>;
+
+export const AllIcons: Story = {
+	args: {
+		size: 24,
+		weight: "regular",
+		className: "fill-primary-brand",
+		alt: "Icons",
+	},
+};
 
 const Template: (iconName: keyof typeof allIcons) => Story = (iconName) => ({
 	args: {
