+++ src/utils/index.ts
@@ -1,3 +1,2 @@
 export * from "./helpers";
-export * from "./mocks";
 export * from "./types";
